+++ connectors/src/api/webhooks/teams/bot_messaging_utils.ts
